add formatter

+++ src/main/java/com/hoanglinh/model/Customer.java
@@ -8,7 +8,7 @@ import javax.persistence.*;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String contact;
@@ -27,18 +27,18 @@ public class Customer {
     public Customer() {
     }
 
-    public Customer(int id, String name, String email, String contact) {
+    public Customer(Long id, String name, String email, String contact) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.contact = contact;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
